rename formatter to tracing
