+++ crates/turborepo-lib/src/lib.rs
@@ -6,11 +6,11 @@ mod commands;
 mod config;
 mod daemon;
 mod execution_state;
-mod formatter;
 pub(crate) mod globwatcher;
 mod package_json;
 mod package_manager;
 mod shim;
+mod tracing;
 mod ui;
 
 use anyhow::Result;
